# Optimize UpdateNodeInfoSnapshot