**What type of PR is this?**

/kind feature

**What this PR does / why we need it**:

Optimizes UpdateNodeInfoSnapshot, which is invoked at the beginning of each scheduling cycle. Profiling 5k cluster benchmarks shows that UpdateNodeInfoSnapshot is one of the most significant overheads. 

Specifically, instead of creating the NodeInfoList every time we schedule a pod, we create it only when a node is added or deleted.

On a test with 5K nodes, 15k existing pods and scheduling 150k pods (which mimics the gce-5000 scalability test),  we get 1.7x performance improvement.

before
```
BenchmarkSchedulingV/5000Nodes/15000Pods-12               150000           8087332 ns/op

BenchmarkScheduling/5000Nodes/1000Pods-12           1000           2233893 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           5557518 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           2441967 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000           4666502 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000           4144397 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000           3531045 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000           7063378 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000           3117295 ns/op
```


after:
```
BenchmarkSchedulingV/5000Nodes/15000Pods-12               150000           4717456 ns/op

BenchmarkScheduling/5000Nodes/1000Pods-12           1000           1675125 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           4741884 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           1750429 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000           3656064 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000           3137824 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000           2657050 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000           6063679 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000           2366480 ns/op

```


**Does this PR introduce a user-facing change?**:
```release-note
NONE
```

This was a result of a joint investigation session with @liu-cong 

/cc @liu-cong