@ahg-g 2019-11-29
> > /hold
> > This is not quite correct.
> 
> Should be correct now, I still need to add unit tests.

/hold cancel

unit tests added.