@liu-cong 2019-11-29
I don't fully agree with how we handle the error when NodeList doesn't agree with NodeMap in the snapshot. But that said, it shouldn't happen assuming implementation is correct, and it's not worse than before, so I will LGTM.

/lgtm