@liu-cong 2019-11-29
/lgtm

Thanks for this big performance improvement!